requirements.txt
discord.py>=2.7.1,<2.8
openai
python-dotenv
requests
ddgs
